Track initial state
Rummikub players start in the initial phase where they have to place a minimal number of points from their own rack before they can make use of sets on the table. When the initial state is set, 'solve' will default to 'solve initial'.

Add 'reset' command to clear the board and set the initial state, and 'initial' to alter the state.