Fix joker count constraint
When repeats == jokers, the joker_constraints list is empty but numbertiles excludes the joker, so the solver would no longer know about the constraint. This isn't a real problem as `tiles <= rack` constraint is stricter as long as we ensure the rack never gains more jokers than are available (which the console tile parser code does).